chore: fix router to recruit  and change directory location
