+++ src/app/_components/lookup-recruitment-form.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import Input from '../shared/input.tsx';
-import { Button } from '../ui/button.tsx';
+import Input from '../../components/shared/input.tsx';
+import { Button } from '../../components/ui/button.tsx';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { useToast } from '../../hooks/use-toast.tsx';
 import { useNavigate } from 'react-router-dom';
