+++ src/app/recruit/[recruitment-id]/_components/recruit-wait/clipboard-section.tsx
@@ -1,4 +1,4 @@
-import Typography from '../../shared/typography';
+import Typography from '../../../../../components/shared/typography';
 
 const ClipboardSection = () => {
   return (
