+++ src/app/recruit/[recruitment-id]/_components/recruit-wait/header-section.tsx
@@ -1,6 +1,6 @@
 import QuestionIcon from '@/assets/icons/question-icon.svg';
-import SvgIcon from '../../shared/svg-icon';
-import Typography from '../../shared/typography';
+import SvgIcon from '../../../../../components/shared/svg-icon';
+import Typography from '../../../../../components/shared/typography';
 
 const HeaderSection = () => {
   return (
